requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
redis==5.0.8
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.3.4
langchain-community==0.3.3
openai==1.51.2
tiktoken==0.7.0
faiss-cpu==1.8.0.post1
beautifulsoup4==4.12.3
websockets==12.0
